refactor(typegen): Move all schema generation stuff into typegen
